requirements.txt
# CORE
pydantic==1.9.0
fastapi==0.73.0
uvicorn==0.17.0.post1

python-dotenv==0.19.2
cryptography==36.0.1

# DB
SQLAlchemy==1.4.31
PyMySQL==1.0.2

# Utils
pydash==5.1.0
dotmap==1.3.23